+++ src/test/davmail/http/TestGetRequest.java
@@ -42,7 +42,7 @@ public class TestGetRequest extends AbstractDavMailTestCase {
 
     public void testRedirect() throws IOException {
         // execute get request, do not follow redirect
-        final String outlookUrl = "https://outlook.office365.com";
+        final String outlookUrl = "https://partner.outlook.cn";
         try (HttpClientAdapter httpClientAdapter = new HttpClientAdapter(outlookUrl)) {
             GetRequest request = new GetRequest(outlookUrl);
             String responseString = httpClientAdapter.executeGetRequest(request);